% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors}
\alias{colors}
\title{Colors}
\format{
A data frame with `r nrow(colors)` rows and `r ncol(colors)` variables.
\describe{
 \item{\code{color_id}}{color ID.}
 \item{\code{name}}{color name.}
 \item{\code{rgb}}{color red, green, and blue (RGB) value.}
 \item{\code{is_trans}}{whether the color is transparent or not.}
}
}
\usage{
colors
}
\description{
Colors in the LEGO Rebrickable data collection
}
\keyword{datasets}
