% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elements}
\alias{elements}
\title{Elements}
\format{
A data frame with `r nrow(elements)` rows and `r ncol(elements)` variables.
\describe{
 \item{\code{element_id}}{element ID.}
 \item{\code{design_id}}{design ID.}
 \item{\code{part_num}}{part number.}
 \item{\code{color_id}}{color ID.}
}
}
\usage{
elements
}
\description{
Elements in the LEGO Rebrickable data collection
}
\keyword{datasets}
