% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventories}
\alias{inventories}
\title{Inventories}
\format{
A data frame with `r nrow(inventories)` rows and `r ncol(inventories)` variables.
\describe{
 \item{\code{inventory_id}}{inventory ID}
 \item{\code{version}}{inventory version number}
 \item{\code{set_num}}{set number}
}
}
\usage{
inventories
}
\description{
Inventories in the LEGO Rebrickable data collection
}
\keyword{datasets}
