% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventory_minifigs}
\alias{inventory_minifigs}
\title{Inventory for mini-figures}
\format{
A data frame with `r nrow(inventory_minifigs)` rows and `r ncol(inventory_minifigs)` variables.
\describe{
 \item{\code{inventory_id}}{inventory ID}
 \item{\code{fig_num}}{figure number}
 \item{\code{quantity}}{number available}
}
}
\usage{
inventory_minifigs
}
\description{
Inventory of mini-figures in the LEGO Rebrickable data collection
}
\keyword{datasets}
