% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventory_parts}
\alias{inventory_parts}
\title{Inventory for parts}
\format{
A data frame with `r nrow(inventory_parts)` rows and `r ncol(inventory_parts)` variables.
\describe{
 \item{\code{inventory_id}}{inventory ID}
 \item{\code{part_num}}{part number.}
 \item{\code{color_id}}{color ID.}
 \item{\code{quantity}}{number available}
 \item{\code{is_spare}}{whether the part is considered a spare or not}
 \item{\code{img_url}}{part image URL}
}
}
\usage{
inventory_parts
}
\description{
Inventory of parts in the LEGO Rebrickable data collection
}
\keyword{datasets}
