% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inventory_sets}
\alias{inventory_sets}
\title{Inventory for sets}
\format{
A data frame with `r nrow(inventory_sets)` rows and `r ncol(inventory_sets)` variables.
\describe{
 \item{\code{inventory_id}}{inventory ID}
 \item{\code{quantity}}{number available}
 \item{\code{set_num}}{set number}
}
}
\usage{
inventory_sets
}
\description{
Inventory of sets in the LEGO Rebrickable data collectoin
}
\keyword{datasets}
