% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minifigs}
\alias{minifigs}
\title{Mini-Figures}
\format{
A data frame with `r nrow(minifigs)` rows and `r ncol(minifigs)` variables.
\describe{
 \item{\code{fig_num}}{figure number}
 \item{\code{name}}{figure name}
 \item{\code{num_parts}}{number of parts in mini-figure}
 \item{\code{img_url}}{mini-figure image URL}
}
}
\usage{
minifigs
}
\description{
Mini-Figures in the LEGO Rebrickable data collection
}
\keyword{datasets}
