% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{part_categories}
\alias{part_categories}
\title{Part Categories}
\format{
A data frame with `r nrow(part_categories)` rows and `r ncol(part_categories)` variables.
\describe{
 \item{\code{part_cat_id}}{part category ID}
 \item{\code{name}}{part category name}
}
}
\usage{
part_categories
}
\description{
Part categories in the LEGO Rebrickable data collection
}
\keyword{datasets}
