% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{part_relationships}
\alias{part_relationships}
\title{Part Relationships}
\format{
A data frame with `r nrow(part_relationships)` rows and `r ncol(part_relationships)` variables.
\describe{
 \item{\code{rel_type}}{Type of relationship. Possible values are `P` (Print), `R` (Pair), `B` (Sub-Part), `M` (Mold), `T` (Pattern), `A` (Alternate)}
 \item{\code{child_part_num}}{child part number}
 \item{\code{parent_part_num}}{parent part number}
}
}
\usage{
part_relationships
}
\description{
Part relationships in the LEGO Rebrickable data collection
}
\keyword{datasets}
