% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parts}
\alias{parts}
\title{Parts}
\format{
A data frame with `r nrow(parts)` rows and `r ncol(parts)` variables.
\describe{
 \item{\code{part_num}}{part number}
 \item{\code{name}}{part name}
 \item{\code{part_cat_id}}{part category ID}
 \item{\code{part_material}}{part material}
}
}
\usage{
parts
}
\description{
Parts in the LEGO Rebricable data collection
}
\keyword{datasets}
