% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sets}
\alias{sets}
\title{Sets}
\format{
A data frame with `r nrow(sets)` rows and `r ncol(sets)` variables.
\describe{
 \item{\code{set_num}}{set number}
 \item{\code{name}}{set name}
 \item{\code{year}}{year set was produced}
 \item{\code{theme_id}}{theme ID}
 \item{\code{num_parts}}{number of parts in set}
 \item{\code{img_url}}{set image URL}
}
}
\usage{
sets
}
\description{
Sets in the LEGO Rebrickable data collection
}
\keyword{datasets}
