% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{themes}
\alias{themes}
\title{Themes}
\format{
A data frame with `r nrow(themes)` rows and `r ncol(themes)` variables.
\describe{
 \item{\code{theme_id}}{theme ID}
 \item{\code{name}}{theme name}
 \item{\code{parent_id}}{parent theme ID}
}
}
\usage{
themes
}
\description{
Themes in the LEGO Rebrickable data collection
}
\keyword{datasets}
